test(reactive): migrate inline tests to integration tests
- Create tests/reactive_tests.rs with 90 tests
- Keep inline tests that access private fields:
  - runtime.rs: test_mark_dirty_duplicate, test_flush_clears_effects
  - tracker.rs: all tests (test internal implementation patterns)
- Migrate tests from: signal.rs, computed.rs, effect.rs, batch.rs,
  context.rs, async_state.rs, mod.rs, runtime.rs (partial)
