+++ src/reactive/batch.rs
@@ -314,384 +314,3 @@ impl Drop for BatchGuard {
         end_batch();
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::sync::atomic::{AtomicBool, AtomicUsize};
-    use std::sync::Arc;
-
-    #[test]
-    fn test_batch_basic() {
-        let counter = Arc::new(AtomicUsize::new(0));
-        let counter_clone = counter.clone();
-
-        batch(|| {
-            queue_update(move || {
-                counter_clone.fetch_add(1, Ordering::SeqCst);
-            });
-        });
-
-        assert_eq!(counter.load(Ordering::SeqCst), 1);
-    }
-
-    #[test]
-    fn test_batch_nested() {
-        let counter = Arc::new(AtomicUsize::new(0));
-
-        batch(|| {
-            let c1 = counter.clone();
-            queue_update(move || {
-                c1.fetch_add(1, Ordering::SeqCst);
-            });
-
-            batch(|| {
-                let c2 = counter.clone();
-                queue_update(move || {
-                    c2.fetch_add(1, Ordering::SeqCst);
-                });
-            });
-
-            // Inner batch shouldn't have flushed yet
-            assert_eq!(counter.load(Ordering::SeqCst), 0);
-        });
-
-        // All updates flushed now
-        assert_eq!(counter.load(Ordering::SeqCst), 2);
-    }
-
-    #[test]
-    fn test_is_batching() {
-        assert!(!is_batching());
-
-        batch(|| {
-            assert!(is_batching());
-        });
-
-        assert!(!is_batching());
-    }
-
-    #[test]
-    fn test_batch_guard() {
-        assert!(!is_batching());
-
-        {
-            let _guard = BatchGuard::new();
-            assert!(is_batching());
-        }
-
-        assert!(!is_batching());
-    }
-
-    #[test]
-    fn test_transaction_commit() {
-        let counter = Arc::new(AtomicUsize::new(0));
-
-        let mut tx = Transaction::new();
-        let c1 = counter.clone();
-        tx.update(move || {
-            c1.fetch_add(1, Ordering::SeqCst);
-        });
-        let c2 = counter.clone();
-        tx.update(move || {
-            c2.fetch_add(1, Ordering::SeqCst);
-        });
-
-        assert_eq!(tx.len(), 2);
-        assert_eq!(counter.load(Ordering::SeqCst), 0);
-
-        tx.commit();
-        assert_eq!(counter.load(Ordering::SeqCst), 2);
-    }
-
-    #[test]
-    fn test_transaction_rollback() {
-        let counter = Arc::new(AtomicUsize::new(0));
-
-        let mut tx = Transaction::new();
-        let c1 = counter.clone();
-        tx.update(move || {
-            c1.fetch_add(1, Ordering::SeqCst);
-        });
-
-        tx.rollback();
-        assert_eq!(counter.load(Ordering::SeqCst), 0);
-    }
-
-    #[test]
-    fn test_flush() {
-        let counter = Arc::new(AtomicUsize::new(0));
-
-        start_batch();
-
-        let c1 = counter.clone();
-        queue_update(move || {
-            c1.fetch_add(1, Ordering::SeqCst);
-        });
-
-        // Force flush while still in batch
-        flush();
-        assert_eq!(counter.load(Ordering::SeqCst), 1);
-
-        end_batch();
-    }
-
-    #[test]
-    fn test_queue_update_no_batch() {
-        let counter = Arc::new(AtomicUsize::new(0));
-        let c1 = counter.clone();
-
-        // Not in batch - should execute immediately
-        queue_update(move || {
-            c1.fetch_add(1, Ordering::SeqCst);
-        });
-
-        assert_eq!(counter.load(Ordering::SeqCst), 1);
-    }
-
-    #[test]
-    fn test_batch_depth() {
-        assert_eq!(batch_depth(), 0);
-
-        start_batch();
-        assert_eq!(batch_depth(), 1);
-
-        start_batch();
-        assert_eq!(batch_depth(), 2);
-
-        end_batch();
-        assert_eq!(batch_depth(), 1);
-
-        end_batch();
-        assert_eq!(batch_depth(), 0);
-    }
-
-    #[test]
-    fn test_batch_count() {
-        let initial = batch_count();
-
-        batch(|| {});
-        assert!(batch_count() > initial);
-
-        batch(|| {
-            batch(|| {});
-        });
-        assert!(batch_count() > initial + 1);
-    }
-
-    #[test]
-    fn test_pending_count() {
-        assert_eq!(pending_count(), 0);
-
-        start_batch();
-
-        queue_update(|| {});
-        assert_eq!(pending_count(), 1);
-
-        queue_update(|| {});
-        assert_eq!(pending_count(), 2);
-
-        end_batch();
-        assert_eq!(pending_count(), 0);
-    }
-
-    #[test]
-    fn test_transaction_is_empty() {
-        let tx = Transaction::new();
-        assert!(tx.is_empty());
-        assert_eq!(tx.len(), 0);
-    }
-
-    #[test]
-    fn test_transaction_len() {
-        let mut tx = Transaction::new();
-        assert_eq!(tx.len(), 0);
-
-        tx.update(|| {});
-        assert_eq!(tx.len(), 1);
-
-        tx.update(|| {});
-        assert_eq!(tx.len(), 2);
-    }
-
-    #[test]
-    fn test_batch_return_value() {
-        let result = batch(|| 42);
-        assert_eq!(result, 42);
-
-        let result = batch(|| "hello".to_string());
-        assert_eq!(result, "hello");
-    }
-
-    #[test]
-    fn test_multiple_flushes() {
-        let counter = Arc::new(AtomicUsize::new(0));
-
-        start_batch();
-
-        let c1 = counter.clone();
-        queue_update(move || {
-            c1.fetch_add(1, Ordering::SeqCst);
-        });
-
-        flush();
-        assert_eq!(counter.load(Ordering::SeqCst), 1);
-
-        let c2 = counter.clone();
-        queue_update(move || {
-            c2.fetch_add(1, Ordering::SeqCst);
-        });
-
-        flush();
-        assert_eq!(counter.load(Ordering::SeqCst), 2);
-
-        end_batch();
-    }
-
-    #[test]
-    fn test_batch_guard_drop() {
-        let counter = Arc::new(AtomicUsize::new(0));
-
-        {
-            let _guard = BatchGuard::new();
-            let c1 = counter.clone();
-            queue_update(move || {
-                c1.fetch_add(1, Ordering::SeqCst);
-            });
-            assert_eq!(counter.load(Ordering::SeqCst), 0);
-            // Guard drops here
-        }
-
-        // After guard drops, updates should be flushed
-        assert_eq!(counter.load(Ordering::SeqCst), 1);
-    }
-
-    #[test]
-    fn test_nested_batch_guard() {
-        let counter = Arc::new(AtomicUsize::new(0));
-
-        {
-            let _outer = BatchGuard::new();
-            assert!(is_batching());
-
-            {
-                let _inner = BatchGuard::new();
-                assert!(is_batching());
-
-                let c1 = counter.clone();
-                queue_update(move || {
-                    c1.fetch_add(1, Ordering::SeqCst);
-                });
-
-                // Inner guard drops - should not flush yet
-            }
-
-            assert_eq!(counter.load(Ordering::SeqCst), 0);
-            assert!(is_batching());
-            // Outer guard drops here
-        }
-
-        assert_eq!(counter.load(Ordering::SeqCst), 1);
-        assert!(!is_batching());
-    }
-
-    #[test]
-    fn test_transaction_default() {
-        let tx = Transaction::default();
-        assert!(tx.is_empty());
-    }
-
-    #[test]
-    fn test_end_batch_without_start() {
-        // Should handle gracefully (saturating_sub)
-        let depth_before = batch_depth();
-        end_batch();
-        let depth_after = batch_depth();
-
-        // Should not go negative
-        assert!(depth_after <= depth_before);
-    }
-
-    #[test]
-    fn test_batch_state_is_thread_local() {
-        // This test verifies the fix for issue #152:
-        // Batch state should be per-thread, not global.
-        // Thread A ending its batch should not affect Thread B's batch state.
-        use std::sync::Barrier;
-        use std::thread;
-
-        let barrier = Arc::new(Barrier::new(2));
-        let thread_b_saw_batching = Arc::new(AtomicBool::new(false));
-
-        let barrier_a = barrier.clone();
-        let barrier_b = barrier.clone();
-        let result = thread_b_saw_batching.clone();
-
-        // Thread A: start batch, wait, then end batch
-        let handle_a = thread::spawn(move || {
-            start_batch();
-            assert!(is_batching());
-            barrier_a.wait(); // Sync point 1: both threads are in their batches
-            barrier_a.wait(); // Sync point 2: wait for B to check state before A ends
-            end_batch();
-            assert!(!is_batching());
-        });
-
-        // Thread B: start batch, wait for A to end, verify B's state is unaffected
-        let handle_b = thread::spawn(move || {
-            start_batch();
-            assert!(is_batching());
-            barrier_b.wait(); // Sync point 1: both threads are in their batches
-            barrier_b.wait(); // Sync point 2: let A proceed to end_batch
-
-            // Small delay to allow Thread A to call end_batch()
-            thread::sleep(std::time::Duration::from_millis(10));
-
-            // Critical check: Thread B should STILL be batching
-            // even though Thread A ended its batch
-            result.store(is_batching(), Ordering::SeqCst);
-            end_batch();
-        });
-
-        handle_a.join().expect("Thread A panicked");
-        handle_b.join().expect("Thread B panicked");
-
-        // Thread B should still have seen itself as batching
-        // despite Thread A ending its batch
-        assert!(
-            thread_b_saw_batching.load(Ordering::SeqCst),
-            "Thread B's batch state was affected by Thread A ending its batch"
-        );
-    }
-
-    #[test]
-    fn test_independent_thread_batches() {
-        // Test that multiple threads can have independent batches
-        use std::thread;
-
-        let counter = Arc::new(AtomicUsize::new(0));
-        let mut handles = vec![];
-
-        for _ in 0..4 {
-            let c = counter.clone();
-            handles.push(thread::spawn(move || {
-                // Each thread has its own batch
-                batch(|| {
-                    assert!(is_batching());
-                    queue_update(move || {
-                        c.fetch_add(1, Ordering::SeqCst);
-                    });
-                    assert!(is_batching());
-                });
-                assert!(!is_batching());
-            }));
-        }
-
-        for handle in handles {
-            handle.join().expect("Thread panicked");
-        }
-
-        // All 4 threads should have executed their updates
-        assert_eq!(counter.load(Ordering::SeqCst), 4);
-    }
-}
