+++ src/reactive/computed.rs
@@ -148,240 +148,3 @@ impl<T: Clone + Send + Sync + 'static> Clone for Computed<T> {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::sync::atomic::AtomicUsize;
-
-    #[test]
-    fn test_computed_basic() {
-        let computed = Computed::new(|| 42);
-        assert_eq!(computed.get(), 42);
-    }
-
-    #[test]
-    fn test_computed_with_closure() {
-        let multiplier = 3;
-        let computed = Computed::new(move || 10 * multiplier);
-        assert_eq!(computed.get(), 30);
-    }
-
-    #[test]
-    fn test_computed_caching() {
-        let call_count = Arc::new(AtomicUsize::new(0));
-        let call_count_clone = call_count.clone();
-
-        let computed = Computed::new(move || {
-            call_count_clone.fetch_add(1, Ordering::SeqCst);
-            42
-        });
-
-        // First call computes
-        assert_eq!(computed.get(), 42);
-        assert_eq!(call_count.load(Ordering::SeqCst), 1);
-
-        // Second call uses cache
-        assert_eq!(computed.get(), 42);
-        assert_eq!(call_count.load(Ordering::SeqCst), 1);
-
-        // After invalidation, recomputes
-        computed.invalidate();
-        assert_eq!(computed.get(), 42);
-        assert_eq!(call_count.load(Ordering::SeqCst), 2);
-    }
-
-    #[test]
-    fn test_computed_dirty_flag() {
-        let computed = Computed::new(|| 1);
-
-        // Initially dirty
-        assert!(computed.is_dirty());
-
-        // After get, not dirty
-        computed.get();
-        assert!(!computed.is_dirty());
-
-        // After invalidate, dirty again
-        computed.invalidate();
-        assert!(computed.is_dirty());
-    }
-
-    #[test]
-    fn test_computed_auto_invalidation() {
-        use crate::reactive::signal;
-
-        let source = signal(10);
-        let compute_count = Arc::new(AtomicUsize::new(0));
-
-        let cc = compute_count.clone();
-        let s = source.clone();
-        let computed = Computed::new(move || {
-            cc.fetch_add(1, Ordering::SeqCst);
-            s.get() * 2
-        });
-
-        // First access computes
-        assert_eq!(computed.get(), 20);
-        assert_eq!(compute_count.load(Ordering::SeqCst), 1);
-
-        // Multiple accesses use cache
-        assert_eq!(computed.get(), 20);
-        assert_eq!(computed.get(), 20);
-        assert_eq!(compute_count.load(Ordering::SeqCst), 1);
-
-        // Update source - should AUTO-INVALIDATE computed!
-        source.set(20);
-
-        // Computed should be dirty now
-        assert!(computed.is_dirty());
-
-        // Next access recomputes automatically
-        assert_eq!(computed.get(), 40);
-        assert_eq!(compute_count.load(Ordering::SeqCst), 2);
-
-        // Cached again
-        assert_eq!(computed.get(), 40);
-        assert_eq!(compute_count.load(Ordering::SeqCst), 2);
-    }
-
-    #[test]
-    fn test_computed_dynamic_dependencies() {
-        use crate::reactive::signal;
-
-        let flag = signal(true);
-        let a = signal(1);
-        let b = signal(2);
-
-        let f = flag.clone();
-        let a_c = a.clone();
-        let b_c = b.clone();
-
-        let computed = Computed::new(move || if f.get() { a_c.get() } else { b_c.get() });
-
-        // Initially: flag=true, depends on a
-        assert_eq!(computed.get(), 1);
-
-        // Change a - should invalidate
-        a.set(10);
-        assert_eq!(computed.get(), 10);
-
-        // Change flag to false - now depends on b
-        flag.set(false);
-        assert_eq!(computed.get(), 2);
-
-        // Change b - should invalidate
-        b.set(20);
-        assert_eq!(computed.get(), 20);
-
-        // Change a - should NOT affect (we don't depend on it anymore)
-        // Note: Due to re-tracking on every get(), this will work correctly
-        a.set(100);
-        assert_eq!(computed.get(), 20); // Still 20, not affected by a
-    }
-
-    #[test]
-    fn test_computed_thread_safety() {
-        use std::thread;
-
-        let computed = Arc::new(Computed::new(|| 42));
-
-        let handles: Vec<_> = (0..4)
-            .map(|_| {
-                let c = computed.clone();
-                thread::spawn(move || {
-                    for _ in 0..100 {
-                        assert_eq!(c.get(), 42);
-                    }
-                })
-            })
-            .collect();
-
-        for h in handles {
-            h.join().unwrap();
-        }
-    }
-
-    #[test]
-    fn test_computed_no_data_race_on_concurrent_recompute() {
-        use std::thread;
-        use std::time::Duration;
-
-        // This test verifies that concurrent calls to get() on a dirty computed
-        // do not cause a data race (double recomputation with inconsistent results).
-        let call_count = Arc::new(AtomicUsize::new(0));
-        let call_count_clone = call_count.clone();
-
-        // Computation that takes some time and increments counter
-        let computed = Arc::new(Computed::new(move || {
-            call_count_clone.fetch_add(1, Ordering::SeqCst);
-            // Small delay to increase chance of race condition
-            thread::sleep(Duration::from_micros(100));
-            42
-        }));
-
-        // Spawn multiple threads that all try to get() simultaneously
-        // when the computed is dirty
-        let handles: Vec<_> = (0..8)
-            .map(|_| {
-                let c = computed.clone();
-                thread::spawn(move || c.get())
-            })
-            .collect();
-
-        let results: Vec<_> = handles.into_iter().map(|h| h.join().unwrap()).collect();
-
-        // All threads should get the same value
-        assert!(results.iter().all(|&v| v == 42));
-
-        // The computation should only have run ONCE despite 8 concurrent callers
-        // (the recompute_lock ensures only one thread recomputes)
-        assert_eq!(
-            call_count.load(Ordering::SeqCst),
-            1,
-            "Computation should run exactly once, but ran {} times",
-            call_count.load(Ordering::SeqCst)
-        );
-    }
-
-    #[test]
-    fn test_computed_recomputes_after_invalidation_with_contention() {
-        use std::thread;
-
-        let call_count = Arc::new(AtomicUsize::new(0));
-        let call_count_clone = call_count.clone();
-
-        let computed = Arc::new(Computed::new(move || {
-            call_count_clone.fetch_add(1, Ordering::SeqCst)
-        }));
-
-        // First get: computes once
-        let v1 = computed.get();
-        assert_eq!(v1, 0);
-        assert_eq!(call_count.load(Ordering::SeqCst), 1);
-
-        // Multiple gets without invalidation: uses cache
-        for _ in 0..10 {
-            assert_eq!(computed.get(), 0);
-        }
-        assert_eq!(call_count.load(Ordering::SeqCst), 1);
-
-        // Invalidate and get from multiple threads
-        computed.invalidate();
-
-        let handles: Vec<_> = (0..4)
-            .map(|_| {
-                let c = computed.clone();
-                thread::spawn(move || c.get())
-            })
-            .collect();
-
-        let results: Vec<_> = handles.into_iter().map(|h| h.join().unwrap()).collect();
-
-        // All should get the new value (1, since call_count was incremented)
-        assert!(results.iter().all(|&v| v == 1));
-
-        // Should have computed exactly twice total (initial + after invalidation)
-        assert_eq!(call_count.load(Ordering::SeqCst), 2);
-    }
-}
