+++ src/reactive/effect.rs
@@ -179,110 +179,3 @@ impl Drop for Effect {
         dispose_subscriber(self.id);
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::sync::atomic::AtomicUsize;
-
-    #[test]
-    fn test_effect_runs_immediately() {
-        let called = Arc::new(AtomicBool::new(false));
-        let called_clone = called.clone();
-
-        let _effect = Effect::new(move || {
-            called_clone.store(true, Ordering::SeqCst);
-        });
-
-        assert!(called.load(Ordering::SeqCst));
-    }
-
-    #[test]
-    fn test_effect_lazy_does_not_run() {
-        let called = Arc::new(AtomicBool::new(false));
-        let called_clone = called.clone();
-
-        let effect = Effect::lazy(move || {
-            called_clone.store(true, Ordering::SeqCst);
-        });
-
-        assert!(!called.load(Ordering::SeqCst));
-
-        effect.run();
-        assert!(called.load(Ordering::SeqCst));
-    }
-
-    #[test]
-    fn test_effect_stop_and_resume() {
-        let count = Arc::new(AtomicUsize::new(0));
-        let count_clone = count.clone();
-
-        let effect = Effect::lazy(move || {
-            count_clone.fetch_add(1, Ordering::SeqCst);
-        });
-
-        effect.run();
-        assert_eq!(count.load(Ordering::SeqCst), 1);
-
-        effect.stop();
-        effect.run();
-        assert_eq!(count.load(Ordering::SeqCst), 1); // Still 1, didn't run
-
-        effect.resume();
-        effect.run();
-        assert_eq!(count.load(Ordering::SeqCst), 2);
-    }
-
-    #[test]
-    fn test_effect_is_active() {
-        let effect = Effect::lazy(|| {});
-
-        assert!(effect.is_active());
-
-        effect.stop();
-        assert!(!effect.is_active());
-
-        effect.resume();
-        assert!(effect.is_active());
-    }
-
-    #[test]
-    fn test_effect_dynamic_dependency_retracking() {
-        use crate::reactive::signal;
-
-        let flag = signal(true);
-        let a = signal(1);
-        let b = signal(2);
-        let result = Arc::new(AtomicUsize::new(0));
-
-        let res = result.clone();
-        let f = flag.clone();
-        let a_c = a.clone();
-        let b_c = b.clone();
-
-        let _effect = Effect::new(move || {
-            // Dynamic dependency: read a or b based on flag
-            let val = if f.get() { a_c.get() } else { b_c.get() };
-            res.store(val as usize, Ordering::SeqCst);
-        });
-
-        // Initially: flag=true, depends on a
-        assert_eq!(result.load(Ordering::SeqCst), 1);
-
-        // Change a - should trigger effect
-        a.set(10);
-        assert_eq!(result.load(Ordering::SeqCst), 10);
-
-        // Change flag - effect should re-track and now depend on b
-        flag.set(false);
-        assert_eq!(result.load(Ordering::SeqCst), 2);
-
-        // Change b - should NOW trigger effect (dynamic re-tracking!)
-        b.set(20);
-        assert_eq!(result.load(Ordering::SeqCst), 20);
-
-        // Change a - should NOT trigger effect anymore (we no longer depend on it)
-        a.set(100);
-        assert_eq!(result.load(Ordering::SeqCst), 20); // Still 20, not affected by a
-    }
-}
